On Insert or Remove, cleanup the txn-queue.
When creating a document in the stash, or moving a document out of the
stash, we can take the opportunity to apply the pull list that we know
about. We do it in memory, rather than having mongo do it, because the
code was already calling "$set" on the txn-queue.