(broken) attempt at synchronizing Find/Remove/Insert.
These don't actually work because while there may not be a *data* race
in memory, they race eachother for things like 'who is the first to
start the transaction'. They also potentially race around 'is the
transaction finished'.

This might be good enough, as we don't really want to try to make it
possible to do things multithreaded, we just don't want the memory
structures to get corrupted.

One caveat is that the TXN isn't actually started until the first
operation, so it may be that we want to support an outer routine calling
StartTransaction, but it forks off a bunch of helper routines that
are all issued inside the transaction. If we want that, then the best
thing is to probably put a Mutex on the 'transaction' object itself. And
then we need to hold the mutex while the query is being performed, so
that we know we really have issued 'startTransaction' on the remote
side.