Cleanup multithreading support.
Clone doesn't copy the session or transaction, that way we don't have to
worry about txn updates, but also we certainly think that Copy() should
have its own transaction identity. And it makes sense that Clone should
too. Especially we don't want Copy during a txn to behave differently
than Copy outside of a transaction.

Get rid of txn.finished. Let the server decide if the transaction is
active.