Read the revnos during the Assert check.
This means we only have to read the docs 1 time. Though ideally we would
push this even higher and not even have to read the docs. (The Assert
*must* be happening in client code anyway, so we just want the txn to
exist for longer than just the Run() operation.)

However, that should bring us to a point where we can make things work.