Improve retry mechanism for transactions when WriteConflict error
This patch aims to fix a juju bug (https://bugs.launchpad.net/juju/+bug/2031631)
by increasing the number of transaction retries in case the returned
mongo error is a WriteConflict error.

This implementation copies the mechanism done in the official golang
mongodb driver, which has a 120 seconds timeout for all transactions and
retries transactions that fail with a transient error until timeout.