feat: close unused sockets
An experiement to close unused sockets after a period of inactivity.
This is a poor mans socket idle timeout. The socket will remain
open even after the socket has been released. During times of high
usage of a session followed by lull periods, the number of unused
sockets can remain very high. This high watermark can exist for
in perpetuity of the application, only a close to the server or a
restart will clear it. This is problematic, as thousands (millions!)
of sockets can remain.

It's not possible to use the pool limit, as load shedding requires
a lot of change to juju itself.

An alternative strategy for the pool limit, would be to close the
socket if the number of live sockets and unused sockets are over a
given number. The cost of creating those new sockets will become
larger, but when the lull period occurs, the number of sockets
will fall to a low watermark.