feat: introduce pool unused limit
When acquiring a socket from the pool, it will also attempt to check
if there are any previously used sockets (unused sockets) and reuse
them for the connection. If there is a burst of socket acquisitions,
then the sockets will be placed back in the unused pool. This creates
a high watermark that never goes back down, at least not without
restarting the application. The existing pool limit could be used
to prevent the excess sockets, though it requires the application to
handle load shedding via the